requirements.txt
# Core dependencies
torch>=2.0.0
transformers>=4.30.0
vllm>=0.2.0
openai>=1.0.0
google-generativeai>=0.8.0
pyyaml>=6.0
nltk>=3.8
scikit-learn>=1.3.0
sentence-transformers>=2.2.0

# Web API dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
python-multipart>=0.0.6

# Utilities
numpy>=1.24.0
pandas>=2.0.0
tqdm>=4.65.0
